refactor: move transaction commit to interpreter
